aws_infra_bot/requirement.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-aws==0.3.0
langchainhub==0.1.21
boto3>=1.34.0
python-dotenv==1.0.1
streamlit==1.32.0
numpy>=1.24.0